Refactor explosions
- Move them to a new file
- Cycle text color instead of background color for sanity
- Clean up some of the expressions
