+++ src/explosion.hpp
@@ -1,0 +1,7 @@
+#pragma once
+
+// Set off an explosion animation at the given sprite coordinate
+void addExplosion(char x, char y);
+
+// Update and draw explosion animations
+void animateExplosions();
